chore: modified couple of things along with size report
