+++ package.json
@@ -15,20 +15,21 @@
     "yarn": "This project is configured to use pnpm"
   },
   "scripts": {
+    "postinstall": "pnpm deps:update",
+    "preinstall": "npx only-allow pnpm",
+    "all": "pnpm -C scripts all",
     "build": "pnpm -r build",
     "clean": "pnpm -r clean && trash .pnpm-store coverage temp && empty-trash",
     "changelog": "auto-changelog -p",
     "create:pkg": "pnpm -C scripts create:pkg",
     "deps:update": "taze -r -w -i -f -l --ignore-paths='third_party/**'",
+    "eslint": "eslint --color --cache --fix --cache-location .eslintcache .",
     "eslint:viewer": "eslint-flat-config-viewer",
     "prepare": "husky install",
-    "preinstall": "npx only-allow pnpm",
-    "postinstall": "pnpm deps:update",
     "knip": "knip --no-gitignore --no-exit-code",
     "lint": "pnpm -C scripts lint",
     "size:report": "pnpm -C scripts size:report",
     "size:limit": "size-limit",
-    "all": "pnpm -C scripts all",
     "secretlint": "secretlint --secretlintignore .gitignore \"**/*\"",
     "spell:check": "cspell .",
     "system:info": "envinfo --system --binaries --browsers",
