+++ packages/cli/package.json
@@ -49,7 +49,6 @@
     "@vitest/coverage-v8": "^0.34.6",
     "trash-cli": "^5.0.0",
     "tsup": "^7.3.0",
-    "vite-tsconfig-paths": "^4.2.2",
     "vitest": "^0.34.6"
   }
 }
