+++ scripts/_templates/package.json.hbs
@@ -40,7 +40,6 @@
     "@vitest/coverage-v8": "^0.34.3",
     "trash-cli": "^5.0.0",
     "tsup": "^7.2.0",
-    "vite-tsconfig-paths": "^4.2.0",
     "vitest": "^0.34.2",
     "wireit": "^0.13.0"
   },
