+++ size-report.md
@@ -9,10 +9,9 @@
 
 | Name   | Size    | Gzip    | Brotli  |
 | ------ | ------- | ------- | ------- |
-| build  | 69.00 B | 88.00 B | 66.00 B |
-| cli    | 66.00 B | 86.00 B | 61.00 B |
-| config | 69.00 B | 88.00 B | 61.00 B |
-| core   | 67.00 B | 86.00 B | 62.00 B |
-| logger | 69.00 B | 87.00 B | 70.00 B |
-
-✅ Ran 1 script and skipped 3 in 0.9s.
+| api    | 62.00 B | 82.00 B | 66.00 B |
+| cli    | 62.00 B | 82.00 B | 66.00 B |
+| config | 62.00 B | 82.00 B | 66.00 B |
+| core   | 62.00 B | 82.00 B | 66.00 B |
+| logger | 62.00 B | 82.00 B | 66.00 B |
+| utils  | 62.00 B | 82.00 B | 66.00 B |
